Supermarket
It's a way to create a printing terminal for receipt.
Here is realisation of Stack data structure. And process of printing is pop(); method.
Sure, it works with DB.
Next I add the addition of new SKU's in DB via this programm.
And, sure, exactly printing with all positions and total summ.